added crutch-patch
It's not good step, but I can't overpass current restriction without losing 'focus' on inputs in settings. And also it will take too much time for making layout using separate methods for each tag.

If this plugin became popular in future, I will rewrite it's layouting logic and even split it for few different files and even add few other small but interesting features. But for now this tradeoff is enough